+++ sdk/metric/internal/aggregate/atomic.go
@@ -228,8 +228,11 @@ type limitedSyncMap[V any] struct {
 	lenMux   sync.Mutex
 }
 
-func (m *limitedSyncMap[V]) LoadOrStoreAttr(fltrAttr attribute.Set, newValue func(attribute.Set) V) V {
-	actual, loaded := m.Load(fltrAttr.Equivalent())
+func (m *limitedSyncMap[V]) LoadOrStoreAttr(
+	lazySet lazyFilteredSet,
+	newValue func(attribute.Set) V,
+) V {
+	actual, loaded := m.Load(lazySet.Distinct())
 	if loaded {
 		return actual.(V)
 	}
@@ -246,13 +249,16 @@ func (m *limitedSyncMap[V]) LoadOrStoreAttr(fltrAttr attribute.Set, newValue fun
 	// re-fetch now that we hold the lock to ensure we don't use the overflow
 	// set unless we are sure the attribute set isn't being written
 	// concurrently.
-	actual, loaded = m.Load(fltrAttr.Equivalent())
+	actual, loaded = m.Load(lazySet.Distinct())
 	if loaded {
 		return actual.(V)
 	}
 
+	var fltrAttr attribute.Set
 	if m.aggLimit > 0 && m.len >= m.aggLimit-1 {
 		fltrAttr = overflowSet
+	} else {
+		fltrAttr = lazySet.Filtered()
 	}
 	actual, loaded = m.LoadOrStore(fltrAttr.Equivalent(), newValue(fltrAttr))
 	if !loaded {